Correct premature match when * is on the same level as a longer match
When rules like:
*.uk
*.sch.uk
are both present, the * in the *.uk rule would cause starfound to be true. This
then caused parentValid to be set when handling found && !n.ExceptRule which
then caused the next iteration to return as a valid TLD index from getTldIndex.
We don't use starfound when we have found a match thus preferring a longer
match.